added WIMN data to the map and html file
Added the WIMN data and styling to the full map.
Had to comment out the health data to run the program. Moved some census styling into the main add to map portion bc I think they need to follow the feature sets that are being created to act on them.